Predicting Name,Age and Count from  API
Based on the input for name,we can predict the age using this API :

https://api.agify.io/?name=${name}

Made use of Axios Library to fetch the API,and useState to predict the age.

We set the initial state to be null,then display all the values of the objects after getting the response,instead of just displaying the age.